pyproject.toml
[tool.poetry]
name = "twofactor"
version = "0.1.0"
description = ""
authors = ["lambdacore <jonathanwh93@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
django-two-factor-auth = "^1.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
